Refactor: Update Pipeline State architecture and DX12 texture upload
- DX12 Restore(WIP)
- Fix PipelineRegistry.h
- Add ComputePipelineState class
- Rename PipelineState to GraphicsPipelineState
- Add CopyDataToTexture2D to handle DX12 texture 256-byte alignment rule

+++ Daydream/Source/Daydream/Graphics/API/DirectX11/D3D11Utility.cpp
@@ -8,7 +8,7 @@ namespace Daydream::GraphicsUtility::DirectX11
 		D3D11_BUFFER_DESC bufferDesc;
 		ZeroMemory(&bufferDesc, sizeof(bufferDesc));
 
-		bufferDesc.ByteWidth = _desc.size;
+		bufferDesc.ByteWidth = (UInt32)_desc.size;
 
 		bufferDesc.Usage = ConvertToD3D11Usage(_desc.memoryUsage);
 		bufferDesc.CPUAccessFlags = ConvertToD3D11CPUAccessFlags(_desc.memoryUsage);
